+++ web/app/page.tsx
@@ -1,14 +1,51 @@
-import Home from './Home'
+'use client'
+import styles from 'styles/Home.module.css'
 
-// async function getPosts() {
-//   const res = await fetch('https://...')
-//   const posts = await res.json()
-//   return posts
-// }
+export default function Home() {
+  return (
+    <div className={styles.container}>
+      <main className={styles.main}>
+        <h1 className={styles.title}>Coming Soon!</h1>
 
-export default function Page() {
-  // Fetch data directly in a Server Component
-  // const recentPosts = await getPosts()
-  // Forward fetched data to your Client Component
-  return <Home />
+        <p className={styles.description}>
+          Refitted: An Android workout tracking app
+          {/* <code className={styles.code}>pages/index.tsx</code> */}
+        </p>
+
+        {/* <div className={styles.grid}>
+      <a href="https://nextjs.org/docs" className={styles.card}>
+        <h2>Documentation &rarr;</h2>
+        <p>Find in-depth information about Next.js features and API.</p>
+      </a>
+
+      <a href="https://nextjs.org/learn" className={styles.card}>
+        <h2>Learn &rarr;</h2>
+        <p>Learn about Next.js in an interactive course with quizzes!</p>
+      </a>
+
+      <a
+        href="https://github.com/vercel/next.js/tree/canary/examples"
+        className={styles.card}
+      >
+        <h2>Examples &rarr;</h2>
+        <p>Discover and deploy boilerplate example Next.js projects.</p>
+      </a>
+
+      <a
+        href="https://vercel.com/new?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+        className={styles.card}
+      >
+        <h2>Deploy &rarr;</h2>
+        <p>
+          Instantly deploy your Next.js site to a public URL with Vercel.
+        </p>
+      </a>
+    </div> */}
+      </main>
+
+      <footer className={styles.footer}>
+        <a href="privacy.html">Privacy Policy</a>
+      </footer>
+    </div>
+  )
 }
